// src/warix.state-action-type.ts
export const enum WarixStateActionType {
    APPLY = '@@apply',
    DELETE = '@@delete',
    LIST_FILTER = '@@list-filter',
    LIST_PUSH = '@@list-push',
    LIST_POP = '@@list-pop',
    LIST_SHIFT = '@@list-shift',
    LIST_UNSHIFT = '@@list-unshift',
    LIST_SPLICE = '@@list-splice',
    LIST_SORT = '@@list-sort',
    PATCH = '@@patch',
    SET = '@@set',
    SET_IN = '@@set-in'
}